Remove XmlBasedImage, xmlbasedimages.py=>xmlutils.py
